Build cli using webpack.
