+++ packages/cli/package.json
@@ -10,8 +10,8 @@
     "dist-types"
   ],
   "scripts": {
-    "build": "tsc --build",
-    "dev": "tsc-watch --build",
+    "build": "webpack --config ../../cli.webpack.config.cjs ",
+    "dev": "webpack --config ../../cli.webpack.config.cjs --watch",
     "build:remove": "rm -rf dist dist-types tsconfig.tsbuildinfo",
     "publish:local": "npm publish --registry=http://localhost:4873",
     "publish:remote": "npm publish --registry=https://registry.npmjs.org",
@@ -21,6 +21,7 @@
   "bin": {
     "psat": "./dist/index.js"
   },
+  "type": "commonjs",
   "repository": {
     "type": "git",
     "url": "https://github.com/GoogleChromeLabs/ps-analysis-tool"
@@ -54,7 +55,8 @@
     "@types/promptly": "^3.0.2",
     "@types/xml2js": "^0.4.14",
     "tsc-watch": "^6.0.4",
-    "typescript": "^5.0.4"
+    "typescript": "^5.0.4",
+    "webpack": "^5.86.0"
   },
   "peerDependencies": {
     "@ps-analysis-tool/cli-dashboard": "*"
