+++ packages/library-detection/src/core/stateProvider/index.ts
@@ -13,31 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * External dependencies.
- */
-import fs from 'fs/promises';
-import path from 'path';
-
-/**
- * Internal dependencies.
- */
-import { CookieDatabase } from '../types';
-
-/**
- * Fetch dictionary from local data folder.
- * @returns {Promise<CookieDatabase>} Open Cookie Data base
- */
-export async function fetchDictionary(): Promise<CookieDatabase> {
-  const data = JSON.parse(
-    await fs.readFile(
-      path.resolve(__dirname, './assets/data/open-cookie-database.json'),
-      {
-        encoding: 'utf8',
-      }
-    )
-  );
-
-  return data;
-}
+export { default as LibraryDetectionProvider } from './libraryDetectionProvider';
+export { default as LibraryDetectionContext } from './context';
+export { default as useLibraryDetectionContext } from './useLibraryDetectionContext';
